requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
numpy==1.25.2
sentence-transformers==2.2.2
google-generativeai==0.2.0
streamlit==1.27.0
pydantic==2.4.2
python-dotenv==1.0.0